% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacon-wrappers.R
\name{MakeBaconDirs}
\alias{MakeBaconDirs}
\title{Create folder and required files for each site}
\usage{
MakeBaconDirs(
  dat = NULL,
  filename,
  path,
  suffix = c("date", "date.time", "none"),
  site.id,
  sample.id,
  age,
  age.err,
  depth
)
}
\arguments{
\item{dat}{A dataframe}

\item{suffix}{Append a suffix to the top-level folder name. Defaults to the
current date. Alternatives are current date and time "date.time" or "none".}

\item{site.id}{Name of variable uniquily identifying sites}

\item{sample.id}{Optional sample.id for Bacon}

\item{age}{Name of variable containing the age estimates}

\item{age.err}{Name of variable containing the 1-sigma age uncertainties}

\item{depth}{Name of variable containing the depth of the samples}
}
\value{
NULL - has side effects of creating a folder structure.
}
\description{
Creates a set of directories and the files required to run
rbacon::Bacon taking data for multiple sites from a dataframe.
}
