% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacon-wrappers.R
\name{RunBaconDirs}
\alias{RunBaconDirs}
\title{Run rbacon::Bacon for data in a set of directories}
\usage{
RunBaconDirs(top.dir.path)
}
\arguments{
\item{top.dir.path}{Top level directory under which are multiple directories
containing input files for Bacon}
}
\value{
NULL - has side effects of writing Bacon output to a folder structure.
}
\description{
Run rbacon::Bacon for data in a set of directories
}
